Bug fixes and Improvements.
Fixed a scaling issue when a Rect is in local coordinate space.

Improvements:
When working with a local space Vector3, the pivot rotation mode in the Unity Editor now changes the orientation of the location handle. Local pivot rotation will allow you to move the Vector3 aligned to the local axis. Global will allow you to move it aligned to the global axis, despite it being a local point. Note that this intentionally makes no different for Vector3's in world space (which is the default).